31th commit: исправил scripts422.sql

+++ scripts422.sql
@@ -1,6 +1,6 @@
 -- создание таблицы машин
 CREATE TABLE auto (
-    id REAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     brand TEXT,
     model TEXT,
     price INTEGER
@@ -8,9 +8,9 @@ CREATE TABLE auto (
 
 -- создание таблицы людей
 CREATE TABLE man (
-    id REAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     name TEXT,
     age INTEGER,
     drive_license BOOLEAN,
-    auto_id REAL REFERENCES auto(id)
+    auto_id INTEGER REFERENCES auto(id)
 );
